Sparse replay storage, BT4-style model arch, ffn_mult float support
Replay buffer:
- Add sparse policy storage: 4 policy fields + legal_mask stored as
  padded-sparse (vals + cols + nnz) in shuffle buffer, ~10x smaller
- Densify on sample — training pipeline sees dense arrays unchanged
- Fix _gather_rows: build prototype from all selected chunks (not just
  first) so optional fields present in any chunk are allocated

Model architecture (BT4-style, requires fresh training):
- Smolgen: rewrite to LC0 BT4 dense-mapping style (per-square compress
  → flatten → MLP → full 64x64 bias per head)
- ValueHead: token_dim 16→32, add mish activation on token projection
- AttentionPolicyHead: policy_dim defaults to embed_dim (was 128)
- Embed: add LayerNorm + gating + post-FFN (BT4 input block)
- ffn_mult: int→float to support fractional multipliers (e.g. 1.5)

Config/CLI:
- pbt2_small.yaml: num_layers 8→9, ffn_mult 2→1.5
- Fix --ffn-mult argparse type int→float in 5 CLI entry points

+++ chess_anti_engine/arena.py
@@ -53,7 +53,7 @@ def main() -> None:
     ap.add_argument("--embed-dim", type=int, default=256)
     ap.add_argument("--num-layers", type=int, default=6)
     ap.add_argument("--num-heads", type=int, default=8)
-    ap.add_argument("--ffn-mult", type=int, default=2)
+    ap.add_argument("--ffn-mult", type=float, default=2)
     ap.add_argument("--use-smolgen", action="store_true")
     ap.add_argument("--use-nla", action="store_true")
     ap.add_argument("--gradient-checkpointing", action="store_true")
@@ -97,7 +97,7 @@ def main() -> None:
             "embed_dim": int(args.embed_dim),
             "num_layers": int(args.num_layers),
             "num_heads": int(args.num_heads),
-            "ffn_mult": int(args.ffn_mult),
+            "ffn_mult": float(args.ffn_mult),
             "use_smolgen": bool(args.use_smolgen),
             "use_nla": bool(args.use_nla),
             "gradient_checkpointing": bool(args.gradient_checkpointing),
@@ -108,7 +108,7 @@ def main() -> None:
         embed_dim=int(mc.get("embed_dim", 256)),
         num_layers=int(mc.get("num_layers", 6)),
         num_heads=int(mc.get("num_heads", 8)),
-        ffn_mult=int(mc.get("ffn_mult", 2)),
+        ffn_mult=float(mc.get("ffn_mult", 2)),
         use_smolgen=bool(mc.get("use_smolgen", True)),
         use_nla=bool(mc.get("use_nla", False)),
         use_gradient_checkpointing=bool(mc.get("gradient_checkpointing", False)),
