+++ chess_anti_engine/bench/play_batch_timing.py
@@ -280,7 +280,7 @@ def main() -> None:
     ap.add_argument("--embed-dim", type=int, default=128)
     ap.add_argument("--num-layers", type=int, default=4)
     ap.add_argument("--num-heads", type=int, default=4)
-    ap.add_argument("--ffn-mult", type=int, default=2)
+    ap.add_argument("--ffn-mult", type=float, default=2)
     ap.add_argument("--no-smolgen", action="store_true")
     ap.add_argument("--use-nla", action="store_true")
 
@@ -307,7 +307,7 @@ def main() -> None:
             embed_dim=int(args.embed_dim),
             num_layers=int(args.num_layers),
             num_heads=int(args.num_heads),
-            ffn_mult=int(args.ffn_mult),
+            ffn_mult=float(args.ffn_mult),
             use_smolgen=not bool(args.no_smolgen),
             use_nla=bool(args.use_nla),
         )
